+++ packages/sources/react/src/components/overlays/VtmnPopover/index.ts
@@ -1,0 +1,2 @@
+export { default as VtmnPopover } from './VtmnPopover';
+export * from './VtmnPopover';