+++ packages/sources/svelte/src/components/selection-controls/VtmnQuantity/test/VtmnQuantity.spec.js
@@ -1,11 +1,7 @@
 import '@testing-library/jest-dom';
 
 import { fireEvent, render } from '@testing-library/svelte';
-import { exec } from 'child_process';
-import { tick } from 'svelte';
-
 import VtmnQuantity from '../VtmnQuantity.svelte';
-const wait = (ms) => new Promise((res) => setTimeout(res, ms));
 
 describe('VtmnQuantity', () => {
   const getQuantity = (container) =>
@@ -182,17 +178,35 @@ describe('VtmnQuantity', () => {
     );
   });
 
-  test('Should trigger blur when blur event on input', async () => {
+  test('Should trigger on:change when change event on input', async () => {
     const { component, container } = render(VtmnQuantity, {
       label: 'unit-test',
       id: 'unit-id',
       error: 'unit test error',
     });
     const input = getInput(container);
-    input.focus();
     const handleEvent = jest.fn();
-    component.$on('blur', handleEvent);
-    input.blur();
+    component.$on('change', handleEvent);
+    await fireEvent.change(input, { target: { value: 10 } });
+    expect(input.value).toBe('10');
     expect(handleEvent).toHaveBeenCalledTimes(1);
+  });
+
+  test('Should trigger on:change when click on add button', async () => {
+    const { container, component } = render(VtmnQuantity, {
+      label: 'unit-test',
+      id: 'unit-id',
+    });
+    const [, addBtn] = getButtons(container);
+    await expectedEventClickOnElement('change', addBtn, component, 1);
+  });
+
+  test('Should trigger on:change when click on subtract button', async () => {
+    const { container, component } = render(VtmnQuantity, {
+      label: 'unit-test',
+      id: 'unit-id',
+    });
+    const [subtractBtn] = getButtons(container);
+    await expectedEventClickOnElement('change', subtractBtn, component, 1);
   });
 });